// src/app/shared/models/cnpj/dados-cnpj.ts
export class DadosCnpjConsultado {
    cnpj: string;
    nome_fantasia: string;
    razao_social: string;
    status: string;
    cnae_principal_descricao: string;
    cnae_principal_codigo: string;
    cep: string;
    data_abertura: string;
    telefone: string;
    email: string;
    tipo_logradouro: string;
    logradouro: string;
    numero: string;
    complemento: string;
    bairro: string;
    cidade: string;
    uf: string;
}